To avoid trunk disruption with a hard-to-build C extension, and until we have a more clear design, moving experimental rating work to a branch.
git-svn-id: https://repos.nextthought.com/svn/nti-svn/NextThoughtPlatform/trunk@8971 3dfb626b-079f-4ab1-a9f6-4654e5377508

+++ bin/InstallDataserver.sh
@@ -43,62 +43,44 @@ else
 	svn co https://svn.nextthought.com/repository/NextThoughtPlatform/trunk $PROJECT_PARENT/NextThoughtPlatform
 fi
 
+export INSTALL_EXTRAS="True"
+
+cd $PROJECT_PARENT/NextThoughtPlatform/nti.dataserver
+
 # NOTE: On AWS, you will need to yum install libxml2-devel and libxslt-devel,
 # and probably add /usr/includ/libxml2/ to CFLAGS
 
-# create a temo work directory
-TMPWK_DIR=`mktemp -d -t tmpwork`
-chmod 777 $TMPWK_DIR
-export PATH=$TMPWK_DIR:$PATH
-	
-# set the "UMFPACK" variable to install scipy
-export UMFPACK="None"
-	
-# make sure we have suitable fortran compiler to install scipy
-pkgs=( g95 gfortran `seq -f "gfortran-mp-4.%g" 4 7` )
-for p in "${pkgs[@]}"
-do
-	cmp=`which $p`
-	if [ -n "$cmp" ]; then
-		ln -s $cmp $TMPWK_DIR/gfortran
-		break
-	elif [ -f "/opt/local/bin/$p" ]; then
-		ln -s "/opt/local/bin/$p" $TMPWK_DIR/gfortran
-		break
-	fi
-done
-
-# install extra packages
-extrap_pkgs=(pyyaml numpy matplotlib scipy pil py)
-for p in "${extrap_pkgs[@]}"
-do
-	echo "Installing $p"
-	pip install -U ${p}
-done
-
-# let's try to use GNU gcc compiler to install scikits.learn
-
-pkgs=( `seq -f "gcc-mp-4.%g" 4 7` )
-for p in "${pkgs[@]}"
-do
-	cmp=`which $p`
-	if [ -n "$cmp" ]; then
-		ln -s $cmp $TMPWK_DIR/gcc
-		break
-	elif [ -f "/opt/local/bin/$p" ]; then
-		ln -s "/opt/local/bin/$p" $TMPWK_DIR/gcc
-		break
-	fi
-done
-
-pip install -U scikits.learn
-
-cd $PROJECT_PARENT/NextThoughtPlatform/nti.dataserver
 pip install -r requirements.txt
-python setup.py develop
 
-# clean
-rm -rf $TMPWK_DIR
+if [ "$INSTALL_EXTRAS" ]; then
+
+	echo "Installing extras"
+
+	if [ -z "$UMFPACK" ]; then
+		export UMFPACK="None"
+	fi
+
+	TMPWK_DIR=`mktemp -d -t tmpwork`
+	chmod 777 $TMPWK_DIR
+	export PATH=$TMPWK_DIR:$PATH
+
+	if [ -f "/opt/local/bin/g95" ]; then
+		ln -s /opt/local/bin/g95 $TMPWK_DIR/g95
+	elif [ -f "/opt/local/bin/gfortran" ]; then
+		ln -s /opt/local/bin/gfortran $TMPWK_DIR/gfortran
+	elif [ -f "/opt/local/bin/gfortran-mp-4.4" ]; then
+		ln -s /opt/local/bin/gfortran-mp-4.4 $TMPWK_DIR/gfortran
+	fi
+
+	pip install pyyaml
+	pip install numpy matplotlib scipy
+	pip install pil
+	pip install py
+
+	rm -rf $TMPWK_DIR
+fi
+
+python setup.py dev
 
 echo "Done."
 echo "Be sure to include the following lines in your .bash_profile:"
